[change] Moved API filter logic to a separate file
